REFACTOR test classes into Excel & Pdf directories
