+++ tests/Excel/ExcelExportFromViewTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Sfneal\ViewExport\Tests;
+namespace Sfneal\ViewExport\Tests\Excel;
 
 use Sfneal\ViewExport\Excel\ExcelExportService;
 use Sfneal\ViewExport\Tests\Assets\ViewModels\TestViewModel;
 
-class ExcelExportFromViewModelTest extends ExcelTestCase
+class ExcelExportFromViewTest extends ExcelTestCase
 {
     /**
      * Setup the test environment.
@@ -16,6 +16,8 @@ class ExcelExportFromViewModelTest extends ExcelTestCase
     {
         parent::setUp();
 
-        $this->renderer = ExcelExportService::fromViewModel(new TestViewModel('table'));
+        $this->renderer = ExcelExportService::fromView(view('table', [
+            'data' => (new TestViewModel())->data(),
+        ]));
     }
 }
